+++ teststudent/src/main/java/org/choongang/global/Updatable.java
@@ -1,0 +1,5 @@
+package org.choongang.global;
+
+public interface Updatable<T> {
+    int update(T form);
+}
