+++ teststudent/src/main/java/org/choongang/student/controllers/RequestMenuNo.java
@@ -1,0 +1,10 @@
+package org.choongang.student.controllers;
+
+
+import lombok.Builder;
+import lombok.Data;
+@Builder
+@Data
+public class RequestMenuNo {
+    private int menuNo;
+}
