+++ teststudent/src/main/java/org/choongang/student/controllers/RequestStundetNo.java
@@ -1,0 +1,11 @@
+package org.choongang.student.controllers;
+
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class RequestStundetNo {
+    private long studentNo;
+}
