+++ teststudent/src/main/java/org/choongang/student/controllers/ScoresController.java
@@ -1,4 +1,4 @@
-package org.choongang.scores.controllers;
+package org.choongang.student.controllers;
 
 import org.choongang.global.AbstractController;
 import org.choongang.global.constants.MainMenu;
