+++ teststudent/src/main/java/org/choongang/student/services/RequestSubject.java
@@ -1,11 +1,11 @@
-package org.choongang.subject.entities;
+package org.choongang.student.services;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
-public class Subject {
+public class RequestSubject {
     private long subCode;
     private String subNm;
     private String teacherNm;
