+++ teststudent/src/main/java/org/choongang/student/services/SearchSubject.java
@@ -1,13 +1,13 @@
-package org.choongang.subject;
+package org.choongang.student.services;
 
 import lombok.Builder;
 import lombok.Data;
-
 @Data
 @Builder
-public class RequestSubject {
+public class SearchSubject {
     private long subCode;
     private String subNm;
     private String teacherNm;
     private String subDiv;
 }
+//실행
