+++ teststudent/src/main/java/org/choongang/student/services/SubjectServiceList.java
@@ -1,19 +1,26 @@
-package org.choongang.subject.services;
+package org.choongang.student.services;
+
 import lombok.RequiredArgsConstructor;
+import org.choongang.global.Listable;
 import org.choongang.global.Service;
-import org.choongang.subject.entities.Subject;
+import org.choongang.student.entities.Subject;
 import org.choongang.student.mapper.SubjectMapper;
 
 import java.util.List;
 
 @RequiredArgsConstructor
-public class SubjectServiceList implements Service<List<Subject>> {
+public class SubjectServiceList implements Service<List<Subject>>, Listable<Object, Subject> {
     private final SubjectMapper mapper;
+
 
     @Override
     public List<Subject> process() {
         System.out.println("**SubjectService-process1()");
+
         return mapper.getSubject();
+
+
 
     }
 }
+
